move middleware into separate module
